+++ frontend/api/internal_api/getInternalUser.ts
@@ -1,7 +1,7 @@
 import { gql } from '@apollo/client';
 
 export const getInternalUser = gql`
-  query GetInternalUser($id: Int!) {
+  query GetInternalUser($id: BigInt!) {
     internalUser(id: $id) {
       id
       email
